Fix adressing issue #32
The bug concerned the line:

if(mExercise.getExerciseSource().equals(ExerciseSource.CUSTOM))

As an NullPointerException has been raised either mExercise or the exercise source was null.
Unfortunatelly I could not reproduce this bug, but at least the app should no longer crash at this point.
I'll close #32 for now, but nevertheless something still seems to be wrong.